requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.11.0
appdirs==1.4.4
astor==0.8.1
astroid==3.2.2
async-timeout==4.0.3
attrs==23.2.0
av==12.3.0
beartype==0.17.2
bitsandbytes==0.43.3
black==24.4.2
blake3==1.0.7
blessed==1.20.0
Brotli==1.0.9
cachetools==6.2.0
calflops==0.0.2
cbor2==5.7.0
cffi==2.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.11.0
contourpy==1.3.0
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==2.20.0
decord==0.6.0
deepspeed==0.14.0
depyf==0.19.0
dill==0.3.6
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring-parser==0.15
einops==0.7.0
email-validator==2.3.0
et_xmlfile==2.0.0
evaluate==0.4.1
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastrlock==0.8.3
filelock==3.19.1
fonttools==4.53.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
fvcore 
gguf==0.17.1
gitdb==4.0.11
GitPython==3.1.42
gmpy2==2.1.2
gpustat==1.1.1
graphviz==0.20.3
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.4
importlib_resources==6.4.5
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
interegular==0.3.3
iopath==0.1.10
isort==5.13.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.4.0
jsonpickle==3.0.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lark==1.2.2
llguidance==0.7.30
llvmlite==0.43.0
lm-format-enforcer==0.11.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mistral_common==1.8.5
mkl==2024.2.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.0.5
multiprocess==0.70.14
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
openai==2.1.0
openai-harmony==0.0.4
openpyxl==3.1.5
outlines_core==0.2.11
packaging==24.0
pandas==2.2.1
partial-json-parser==0.2.1.1.post6
pathspec==0.12.1
patsy==1.0.1
pillow==11.3.0
platformdirs==4.2.2
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
protobuf==4.25.3
psutil==5.9.8
ptflops==0.7.4
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.10
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.17.2
pylint==3.2.3
pynvml==11.5.0
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
ray==2.49.2
referencing==0.36.2
regex==2023.12.25
requests==2.32.3
responses==0.18.0
rich==13.7.1
rich-toolkit==0.15.1
rignore==0.7.0
rotary-embedding-torch==0.5.3
rouge_score==0.1.2
rpds-py==0.27.1
sacred==0.8.5
safetensors==0.6.2
SciencePlots==2.1.1
scikit-learn==1.6.0
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.39.0
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.7.1
smmap==5.0.1
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
speculative-decoding==0.1.2
starlette==0.48.0
statsmodels==0.14.4
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
termcolor==3.1.0
threadpoolctl==3.5.0
tiktoken==0.11.0
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
transformers==4.44.0
triton==2.3.1
trl==0.8.1
typer==0.19.2
typing-inspection==0.4.2
tyro==0.7.3
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.37.0
uvloop==0.21.0
wandb==0.16.4
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.16.0
xgrammar==0.1.25
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.20.1
